finetune/requirements.txt
transformers
peft
accelerate
trl
bitsandbytes
datasets
torch
torchvision
torchaudio
pillow
PyYAML
deepspeed
python-json-logger
# liger-kernel
